chore: move envs into subfolder
