+++ neurogym/envs/native/economicdecisionmaking.py
