+++ neurogym/envs/native/intervaldiscrimination.py
