+++ tests/envs/test_annubes.py
@@ -3,7 +3,7 @@ from collections.abc import Callable
 import numpy as np
 import pytest
 
-from neurogym.envs.annubes import AnnubesEnv
+from neurogym.envs.native.annubes import AnnubesEnv
 
 RND_SEED = 42
 FIX_INTENSITY = 0.1
